I have no idea if this is (a) idiomatic go, and (b) a realistic thing to care about for the usage of this api, I'm just amazed I managed to get the tests to pass.

```
Before:
goos: darwin
goarch: amd64
pkg: github.com/kiambogo/go-hypercore/mempager
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
Benchmark_PagerGrowPages100-12       	  750114	      1357 ns/op	    4152 B/op	      12 allocs/op
Benchmark_PagerGrowPages100000-12    	     784	   1558879 ns/op	 4656489 B/op	      34 allocs/op


After:
goos: darwin
goarch: amd64
pkg: github.com/kiambogo/go-hypercore/mempager
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
Benchmark_PagerGrowPages100-12       	 1000000	      1006 ns/op	    3936 B/op	       8 allocs/op
Benchmark_PagerGrowPages100000-12    	    2830	    510968 ns/op	 1607811 B/op	       8 allocs/op
```